emotorad/mqtt-publisher/requirements.txt
paho-mqtt==1.6.1
redis==3.5.3
